move issue tests into their own files
