+++ test/fake-timers-test.js
@@ -1,277 +1,25 @@
 "use strict";
 
-describe("issue #2449: permanent loss of native functions", function () {
-    it("should not fake faked timers", function () {
-        const currentTime = new Date().getTime();
-        const date1 = new Date("2015-09-25");
-        const date2 = new Date("2015-09-26");
-        let clock = FakeTimers.install({ now: date1 });
-        assert.same(clock.now, date1.getTime());
-        assert.same(new Date().getTime(), 1443139200000);
-        assert.exception(function () {
-            FakeTimers.install({ now: date2 });
-        });
-        clock.uninstall();
-        clock = FakeTimers.install({ now: date2 });
-        assert.same(clock.now, date2.getTime());
-        clock.uninstall();
-        assert.greater(new Date().getTime(), currentTime, true);
-    });
-
-    it("should not fake faked timers on a custom target", function () {
-        const setTimeoutFake = sinon.fake();
-        const context = {
-            Date: Date,
-            setTimeout: setTimeoutFake,
-            clearTimeout: sinon.fake(),
-        };
-        let clock = FakeTimers.withGlobal(context).install();
-        assert.exception(function () {
-            clock = FakeTimers.withGlobal(context).install();
-        });
-        clock.uninstall();
-
-        // After uninstaling we should be able to install without issue
-        clock = FakeTimers.withGlobal(context).install();
-        clock.uninstall();
-    });
-
-    it("should not allow a fake on a custom target if the global is faked and the context inherited from the global", function () {
-        const globalClock = FakeTimers.install();
-        assert.equals(new Date().getTime(), 0);
-        const setTimeoutFake = sinon.fake();
-        const context = {
-            Date: Date,
-            setTimeout: setTimeoutFake,
-            clearTimeout: sinon.fake(),
-        };
-        assert.equals(new context.Date().getTime(), 0);
-        assert.exception(function () {
-            FakeTimers.withGlobal(context).install();
-        });
-
-        globalClock.uninstall();
-        refute.equals(new Date().getTime(), 0);
-    });
-
-    it("should allow a fake on the global if a fake on a customer target is already defined", function () {
-        const setTimeoutFake = sinon.fake();
-        const context = {
-            Date: Date,
-            setTimeout: setTimeoutFake,
-            clearTimeout: sinon.fake(),
-        };
-        const clock = FakeTimers.withGlobal(context).install();
-        assert.equals(new context.Date().getTime(), 0);
-        refute.equals(new Date().getTime(), 0);
-        const globalClock = FakeTimers.install();
-        assert.equals(new Date().getTime(), 0);
-
-        globalClock.uninstall();
-        refute.equals(new Date().getTime(), 0);
-        assert.equals(new context.Date().getTime(), 0);
-        clock.uninstall();
-        refute.equals(new Date().getTime(), 0);
-        refute.equals(new context.Date().getTime(), 0);
-    });
-});
-describe("issue #59", function () {
-    it("should install and uninstall the clock on a custom target", function () {
-        const setTimeoutFake = sinon.fake();
-        const context = {
-            Date: Date,
-            setTimeout: setTimeoutFake,
-            clearTimeout: sinon.fake(),
-        };
-        const clock = FakeTimers.withGlobal(context).install();
-        assert.equals(setTimeoutFake.callCount, 1);
-        clock.setTimeout(NOOP, 0);
-        assert.equals(setTimeoutFake.callCount, 1);
-        // this would throw an error before issue #59 was fixed
-        clock.uninstall();
-    });
-});
-describe("issue #73", function () {
-    it("should install with date object", function () {
-        const date = new Date("2015-09-25");
-        const clock = FakeTimers.install({ now: date });
-        assert.same(clock.now, 1443139200000);
-        clock.uninstall();
-    });
-});
-
-describe("issue #67", function () {
-    // see https://nodejs.org/api/timers.html
-    it("should overflow to 1 on very big timeouts", function () {
-        const clock = FakeTimers.install();
-        const stub1 = sinon.stub();
-        const stub2 = sinon.stub();
-
-        clock.setTimeout(stub1, 100);
-        clock.setTimeout(stub2, 214748334700); //should be called after 1 tick
-
-        clock.tick(1);
-        assert(stub2.called);
-        assert.isFalse(stub1.called);
-
-        clock.tick(99);
-        assert(stub1.called);
-        assert(stub2.called);
-
-        clock.uninstall();
-    });
-
-    it("should overflow to interval 1 on very big timeouts", function () {
-        const clock = FakeTimers.install();
-        const stub = sinon.stub();
-
-        clock.setInterval(stub, 214748334700);
-        clock.tick(3);
-        assert(stub.calledThrice);
-
-        clock.uninstall();
-    });
-
-    it("should execute setTimeout smaller than 1", function () {
-        const clock = FakeTimers.install();
-        const stub1 = sinon.stub();
-
-        clock.setTimeout(stub1, 0.5);
-        clock.tick(1);
-        assert(stub1.calledOnce);
-
-        clock.uninstall();
-    });
-
-    it("executes setTimeout with negative duration as if it has zero delay", function () {
-        const clock = FakeTimers.install();
-        const stub1 = sinon.stub();
-
-        clock.setTimeout(stub1, -10);
-        clock.tick(1);
-        assert(stub1.calledOnce);
-
-        clock.uninstall();
-    });
-});
-
-describe("issue sinon#1852", function () {
-    it("throws when creating a clock and global has no Date", function () {
-        const clock = FakeTimers.withGlobal({
-            setTimeout: function () {},
-            clearTimeout: function () {},
-        });
-        assert.exception(function () {
-            clock.createClock();
-        });
-        assert.exception(function () {
-            clock.install();
-        });
-    });
-});
-
-describe("issue #207 - nanosecond round-off errors on high-res timer", function () {
-    let clock;
-
-    afterEach(function () {
-        clock.uninstall();
-    });
-
-    if (hrtimePresent) {
-        it("should not round off nanosecond arithmetic on hrtime - case 1", function () {
-            clock = FakeTimers.install();
-
-            clock.tick(1022.7791);
-
-            const nanos = clock.hrtime([0, 2 * 1e7])[1];
-            assert.equals(nanos, 2779100);
-        });
-
-        it("should not round off nanosecond arithmetic on hrtime - case 2", function () {
-            clock = FakeTimers.install({
-                now: new Date("2018-09-12T08:58:33.742000000Z").getTime(),
-                toFake: ["hrtime"],
-            });
-            const start = clock.hrtime();
-            clock.tick(123.493);
-
-            const nanos = clock.hrtime(start)[1];
-            assert.equals(nanos, 123493000);
-        });
-
-        it("should truncate sub-nanosecond ticks", function () {
-            clock = FakeTimers.install();
-            clock.tick(0.123456789);
-
-            const nanos = clock.hrtime()[1];
-            assert.equals(nanos, 123456);
-        });
-    }
-
-    it("should always set 'now' to an integer value when ticking with sub-millisecond precision", function () {
-        clock = FakeTimers.install();
-        clock.tick(2.993);
-
-        assert.equals(clock.now, 2);
-    });
-
-    it("should adjust adjust the 'now' value when the nano-remainder overflows", function () {
-        clock = FakeTimers.install();
-        clock.tick(0.993);
-        clock.tick(0.5);
-
-        assert.equals(clock.now, 1);
-    });
-
-    it("should floor negative now values", function () {
-        clock = FakeTimers.install({ now: -1.2 });
-
-        assert.equals(clock.now, -2);
-    });
-
-    it("should floor start times", function () {
-        clock = FakeTimers.install({ now: 1.2 });
-        assert.equals(clock.now, 1);
-    });
-
-    it("should floor negative start times", function () {
-        clock = FakeTimers.install({ now: -1.2 });
-        assert.equals(clock.now, -2);
-    });
-
-    it("should handle ticks on the negative side of the Epoch", function () {
-        clock = FakeTimers.install({ now: -2 });
-        clock.tick(0.8); // -1.2
-        clock.tick(0.5); // -0.7
-
-        assert.equals(clock.now, -1);
-    });
-
-    it("should handle multiple non-integer ticks", function () {
-        clock = FakeTimers.install({ now: -2 });
-        clock.tick(1.1); // -0.9
-        clock.tick(0.5);
-        clock.tick(0.5); // 0.1
-
-        assert.equals(clock.now, 0);
-    });
-});
-
-describe("issue #sinonjs/2086 - don't install setImmediate in unsupported environment", function () {
-    let clock;
-
-    if (typeof setImmediate === "undefined") {
-        afterEach(function () {
-            clock.uninstall();
-        });
-
-        it("should not install setImmediate", function () {
-            clock = FakeTimers.install();
-
-            assert.isUndefined(global.setImmediate);
-        });
-    }
-});
+const {
+    addTimerReturnsObject,
+    assert,
+    FakeTimers,
+    GlobalDate,
+    globalObject,
+    hrtimeBigintPresent,
+    hrtimePresent,
+    nextTickPresent,
+    NOOP,
+    performanceMarkPresent,
+    performanceNowPresent,
+    promisePresent,
+    queueMicrotaskPresent,
+    refute,
+    setImmediatePresent,
+    sinon,
+    utilPromisify,
+    utilPromisifyAvailable,
+} = require("./helpers/setup-tests");
 
 describe("FakeTimers", function () {
     describe("setTimeout", function () {
@@ -4833,123 +4581,6 @@ describe("FakeTimers", function () {
             this.clock.runAll();
 
             assert.isFalse(stub.called);
-        });
-    });
-});
-
-describe("#276 - remove config.target", function () {
-    it("should throw on using `config.target`", function () {
-        assert.exception(function () {
-            FakeTimers.install({ target: {} });
-        }, /config.target is no longer supported/);
-    });
-});
-
-describe("issue #315 - praseInt if delay is not a number", function () {
-    it("should successfully execute the timer", function () {
-        const clock = FakeTimers.install();
-        const stub1 = sinon.stub();
-
-        clock.setTimeout(stub1, "1");
-        clock.tick(1);
-        assert(stub1.calledOnce);
-
-        clock.uninstall();
-    });
-});
-
-describe("#368 - timeout.refresh setTimeout arguments", function () {
-    before(function () {
-        if (!addTimerReturnsObject) {
-            this.skip();
-        }
-    });
-
-    it("should forward arguments passed to setTimeout", function () {
-        const clock = FakeTimers.install();
-        const stub = sinon.stub();
-
-        const t = setTimeout(stub, 1000, "test");
-        clock.tick(1000);
-        t.refresh();
-        clock.tick(1000);
-        assert.calledTwice(stub);
-        assert.alwaysCalledWith(stub, "test");
-        clock.uninstall();
-    });
-});
-
-describe("#187 - Support timeout.refresh in node environments", function () {
-    it("calls the stub again after refreshing the timeout", function () {
-        const clock = FakeTimers.install();
-        const stub = sinon.stub();
-
-        if (typeof setTimeout(NOOP, 0) === "object") {
-            const t = setTimeout(stub, 1000);
-            clock.tick(1000);
-            t.refresh();
-            clock.tick(1000);
-            assert(stub.calledTwice);
-        }
-        clock.uninstall();
-    });
-
-    it("only calls stub once if not fired at time of refresh", function () {
-        const clock = FakeTimers.install();
-        const stub = sinon.stub();
-
-        if (typeof setTimeout(NOOP, 0) === "object") {
-            const t = setTimeout(stub, 1000);
-            clock.tick(999);
-            assert(stub.notCalled);
-            t.refresh();
-            clock.tick(999);
-            assert(stub.notCalled);
-            clock.tick(1);
-            assert(stub.calledOnce);
-        }
-        clock.uninstall();
-    });
-});
-
-describe("#347 - Support util.promisify once installed", function () {
-    before(function () {
-        if (!utilPromisifyAvailable) {
-            this.skip();
-        }
-    });
-
-    beforeEach(function () {
-        this.clock = FakeTimers.install();
-    });
-
-    afterEach(function () {
-        this.clock.uninstall();
-    });
-
-    it("setTimeout", function () {
-        let resolved = false;
-        utilPromisify(global.setTimeout)(100).then(function () {
-            resolved = true;
-        });
-
-        return this.clock.tickAsync(100).then(function () {
-            assert.isTrue(resolved);
-        });
-    });
-
-    it("setImmediate", function () {
-        if (!setImmediatePresent) {
-            this.skip();
-        }
-
-        let resolved = false;
-        utilPromisify(global.setImmediate)().then(function () {
-            resolved = true;
-        });
-
-        return this.clock.tickAsync(0).then(function () {
-            assert.isTrue(resolved);
         });
     });
 });
