+++ test/issue-276-test.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const { FakeTimers, assert } = require("./helpers/setup-tests");
+
+describe("#276 - remove config.target", function () {
+    it("should throw on using `config.target`", function () {
+        assert.exception(function () {
+            FakeTimers.install({ target: {} });
+        }, /config.target is no longer supported/);
+    });
+});
